+++ helpers/python.ts
@@ -167,7 +167,7 @@ const getAdditionalDependencies = (
         version: "0.1.6",
       });
       break;
-    case "llmhub":
+    case "t-systems":
       dependencies.push({
         name: "llama-index-agent-openai",
         version: "0.2.2",